Cease assuming size of int

+++ Software/tcc_li_fi/encoding_8b10b.h
@@ -5,14 +5,14 @@
 #define INITIAL_RD_VALUE -1
 
 // Running Disparity
-static char rd = INITIAL_RD_VALUE;
+static signed char rd = INITIAL_RD_VALUE;
 
 // Encoding functions
 void setEncodingRD(char rdValue);
 void setup8B10B();
-unsigned int encode8B10B(unsigned char data);
-unsigned char decode8B10B(unsigned int data);
-unsigned int encode10bControl(unsigned int controlNumber);
-unsigned int decode10bControl(unsigned int encodedControl);
+uint16_t encode8B10B(unsigned char data);
+unsigned char decode8B10B(uint16_t data);
+uint16_t encode10bControl(uint16_t controlNumber);
+uint16_t decode10bControl(uint16_t encodedControl);
 
 #endif
